applied patches against warnings, thanks a lot Ole K. http://code.google.com/p/bullet/issues/detail?id=165 http://code.google.com/p/bullet/issues/detail?id=164 http://code.google.com/p/bullet/issues/detail?id=166 http://code.google.com/p/bullet/issues/detail?id=167

+++ src/BulletCollision/CollisionShapes/btTriangleMesh.cpp
@@ -111,9 +111,9 @@ int	btTriangleMesh::findOrAddVertex(const btVector3& vertex, bool removeDuplicat
 				}
 			}
 	}
-		m_3componentVertices.push_back(vertex.getX());
-		m_3componentVertices.push_back(vertex.getY());
-		m_3componentVertices.push_back(vertex.getZ());
+		m_3componentVertices.push_back((float)vertex.getX());
+		m_3componentVertices.push_back((float)vertex.getY());
+		m_3componentVertices.push_back((float)vertex.getZ());
 		m_indexedMeshes[0].m_numVertices++;
 		m_indexedMeshes[0].m_vertexBase = (unsigned char*)&m_3componentVertices[0];
 		return (m_3componentVertices.size()/3)-1;
